Make GoalSelectorDropdown get data from store
