+++ src/components/GoalView/GoalSelector/GoalSelectorDropdown/index.tsx
@@ -1,0 +1,13 @@
+import { GoalSelectorDropdownState } from "../../../../types/goals";
+import { GoalSelectorDropdown } from "./GoalSelectorDropdown";
+
+import { connect } from "react-redux";
+import { StoreState } from "../../../../types";
+
+export function mapStateToProps(state: StoreState): GoalSelectorDropdownState {
+  return {
+    possibleGoals: state.goalsState.all
+  };
+}
+
+export default connect(mapStateToProps)(GoalSelectorDropdown);
